chronoGUI: Change unsigned int to long int to accomodate for a bigger amount of days.

+++ src/chronoGUI.cpp
@@ -347,12 +347,12 @@ void setChronoGUI(Menu* menu, chronometer* tchrono) {
   strcpy(sel.subtitle, "Days");
   sel.value = 0;
   sel.min = 0;
-  sel.max = -1; // no limit. long long it is big enough to accomodate a chronometer with a duration of over 2 million days.
+  sel.max = -1; // no limit. long long int is big enough to accomodate a chronometer with a duration of over 2 million days.
   sel.cycle = 0;
   sel.type = SELECTORTYPE_NORMAL;
   res = doSelector(&sel);
   if (res == SELECTOR_RETURN_EXIT) return;
-  unsigned int days = sel.value;
+  long int days = sel.value;
   
   strcpy(sel.title, "Set timer");
   strcpy(sel.subtitle, "Hours");
